Refactored AS::DescendantsTracker test cases so they can be tested without AS::Dependencies.
